cmd/common: move `version` and `sync` command
The version and sync commands will be used in multiple binaries. In order to reduce code,
this moves the commands to the common package.
