+++ README.md
@@ -106,7 +106,7 @@ Available Commands:
   prometheus  Expose metrics using prometheus
   script      Run a bpftrace-compatible scripts
   snapshot    Take a snapshot of a subsystem and print it
-  sync        Synchronize gadget information with your cluster
+  sync        Synchronize gadget information with server
   top         Gather, sort and periodically report events according to a given criteria
   trace       Trace and print system events
   traceloop   Get strace-like logs of a container from the past
