+++ cmd/kubectl-gadget/deploy.go
@@ -49,6 +49,7 @@ import (
 	watchtools "k8s.io/client-go/tools/watch"
 	"sigs.k8s.io/yaml"
 
+	"github.com/inspektor-gadget/inspektor-gadget/cmd/common"
 	commonutils "github.com/inspektor-gadget/inspektor-gadget/cmd/common/utils"
 	"github.com/inspektor-gadget/inspektor-gadget/cmd/kubectl-gadget/utils"
 	"github.com/inspektor-gadget/inspektor-gadget/pkg/k8sutil"
@@ -448,7 +449,7 @@ func runDeploy(cmd *cobra.Command, args []string) error {
 				case "GADGET_IMAGE":
 					gadgetContainer.Env[i].Value = image
 				case "INSPEKTOR_GADGET_VERSION":
-					gadgetContainer.Env[i].Value = version
+					gadgetContainer.Env[i].Value = common.Version()
 				case "INSPEKTOR_GADGET_OPTION_HOOK_MODE":
 					gadgetContainer.Env[i].Value = hookMode
 				case "INSPEKTOR_GADGET_OPTION_FALLBACK_POD_INFORMER":
