+++ cmd/kubectl-gadget/main.go
@@ -81,13 +81,7 @@ func main() {
 	// Advise category is still being handled by CRs for now
 	rootCmd.AddCommand(advise.NewAdviseCmd())
 
-	rootCmd.AddCommand(&cobra.Command{
-		Use:   "sync",
-		Short: "Synchronize gadget information with your cluster",
-		RunE: func(cmd *cobra.Command, args []string) error {
-			return runtime.UpdateDeployInfo()
-		},
-	})
+	rootCmd.AddCommand(common.NewSyncCommand(runtime))
 
 	if err := rootCmd.Execute(); err != nil {
 		os.Exit(1)
